fix(mobile): review fixes — inline regex indices, scoped retry, transport-close handling
Review findings (Greptile P1 ×2, CodeRabbit critical + majors):

- markdown: image alternative has 2 capture groups, not 3 — inline
  segment indices were shifted by one, so links dropped, images crashed
  segmentKey, and bold/strike/code/math all misrendered. Indices fixed;
  parser extracted to pure markdown-parser.ts with a 12-case vitest
  suite pinning one-segment-per-construct (new test target).
- retry: bind retry to the failed assistant message id (kept in sync
  when the server assigns the real id); read finalized messages from
  the React Query cache when settlement cleared the transient store —
  no more duplicated user turns.
- streaming: a transport close before the backend done event now
  settles the turn as an error (partial text kept, retryable) instead
  of silently persisting a truncated answer as complete.
- sse-client: disarm the stall watchdog before scheduling reconnects so
  a stale timer can't cancel retry backoff.
- conversations: rename/delete/star only touch the cache when the API
  resolves true (these helpers return false on failure instead of
  throwing); chat-history uses the hook mutators directly.
- table keys derived from precomputed positional ids — removes the
  biome-ignore suppressions (suppression-baseline check).
- tsconfig: exclude *.test.ts from the app type-check (vitest pulls
  Node fetch types that clash with Expo's DOM types).
