+++ apps/mobile/src/features/chat/hooks/use-conversations.ts
@@ -14,10 +14,10 @@ interface UseConversationsReturn {
   isLoading: boolean;
   error: string | null;
   refetch: () => Promise<void>;
-  deleteConversation: (id: string) => Promise<void>;
-  renameConversation: (id: string, title: string) => Promise<void>;
-  starConversation: (id: string) => Promise<void>;
-  unstarConversation: (id: string) => Promise<void>;
+  deleteConversation: (id: string) => Promise<boolean>;
+  renameConversation: (id: string, title: string) => Promise<boolean>;
+  starConversation: (id: string) => Promise<boolean>;
+  unstarConversation: (id: string) => Promise<boolean>;
 }
 
 export function useConversations(): UseConversationsReturn {
@@ -26,28 +26,41 @@ export function useConversations(): UseConversationsReturn {
   const { updateConversationInCache, removeConversationFromCache } =
     useChatQueryClient();
 
-  const deleteConversation = async (id: string) => {
-    await chatApi.deleteConversation(id);
+  // These API helpers resolve false on failure instead of throwing, so the
+  // result must be checked BEFORE touching the optimistic cache.
+  const deleteConversation = async (id: string): Promise<boolean> => {
+    const ok = await chatApi.deleteConversation(id);
+    if (!ok) return false;
     removeConversationFromCache(id);
     queryClient.invalidateQueries({ queryKey: chatKeys.conversations() });
+    return true;
   };
 
-  const renameConversation = async (id: string, title: string) => {
-    await chatApi.renameConversation(id, title);
+  const renameConversation = async (
+    id: string,
+    title: string,
+  ): Promise<boolean> => {
+    const ok = await chatApi.renameConversation(id, title);
+    if (!ok) return false;
     updateConversationInCache(id, { title });
     queryClient.invalidateQueries({ queryKey: chatKeys.conversations() });
+    return true;
   };
 
-  const starConversation = async (id: string) => {
-    await chatApi.toggleStarConversation(id, true);
+  const starConversation = async (id: string): Promise<boolean> => {
+    const ok = await chatApi.toggleStarConversation(id, true);
+    if (!ok) return false;
     updateConversationInCache(id, { is_starred: true });
     queryClient.invalidateQueries({ queryKey: chatKeys.conversations() });
+    return true;
   };
 
-  const unstarConversation = async (id: string) => {
-    await chatApi.toggleStarConversation(id, false);
+  const unstarConversation = async (id: string): Promise<boolean> => {
+    const ok = await chatApi.toggleStarConversation(id, false);
+    if (!ok) return false;
     updateConversationInCache(id, { is_starred: false });
     queryClient.invalidateQueries({ queryKey: chatKeys.conversations() });
+    return true;
   };
 
   return {
