+++ apps/mobile/tsconfig.json
@@ -16,5 +16,9 @@
     ".expo/types/**/*.ts",
     "expo-env.d.ts",
     "src/uniwind-types.d.ts"
+  ],
+  "exclude": [
+    "src/**/*.test.ts",
+    "vitest.config.ts"
   ]
 }
